Server runs, able to deploy to GCP, and added  Map API test samples.
So I found out that after running this app on my other computer, it didn't want to work because I forgot to include another file. Now this app will work anywhere with that missing file. I also found out how to deploy on GCP, really really easily actually. And I added a few test samples for the Maps API with instructions on how to use.